Not Executable - DX12 Add D3D12MemoryAllcator(Buffer, Texture) - Complete Change Renderer Function Params(shared -> raw ptr) - Add UploadBufferPool - Fix Sampler Registry

+++ Daydream/Source/Daydream/Graphics/Pools/TexturePool/Texture2DPoolHandle.h
@@ -22,7 +22,7 @@ namespace Daydream
 		Texture2DPoolHandle(Texture2DPoolHandle&&) noexcept = default;
 		Texture2DPoolHandle& operator=(Texture2DPoolHandle&&) noexcept = default;
 	private:
-		UInt32 lastUsedLoop = 0;
+		UInt64 lastUsedLoop = 0;
 		Shared<Texture2D> texture;
 		Shared<TextureView> renderTargetView;
 		Shared<TextureView> depthStencilView;
